@@ -13,26 +13,31 @@
     "lint": "prettier-standard 'src/**/*.js' --lint"
   },
   "dependencies": {
-    "@codemirror/lang-javascript": "^0.20.1",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@reduxjs/toolkit": "^1.8.2",
-    "@uiw/react-codemirror": "^4.7.0",
-    "firebase": "^7.19.0",
+    "@codemirror/lang-javascript": "0.20.1",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.14.16",
+    "@mui/material": "^5.14.16",
+    "@mui/styles": "^5.14.16",
+    "@mui/x-tree-view": "^6.17.0",
+    "@reduxjs/toolkit": "^1.9.7",
+    "@uiw/react-codemirror": "4.7.0",
+    "firebase": "^10.5.2",
     "lodash": "^4.17.20",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-redux": "^7.2.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "redux-saga": "^1.1.3"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.3",
+    "react-router-dom": "^6.18.0",
+    "react-scripts": "^5.0.1",
+    "redux-saga": "^1.2.3"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^8.1.3",
-    "eslint": "^7.32.0",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@testing-library/jest-dom": "^6.1.4",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.5.1",
+    "eslint": "7.32.0",
+    "postcss": "^8.4.31",
     "prettier-standard": "^16.4.1"
   },
   "browserslist": {
