@@ -1,5 +1,5 @@
 import React, { memo } from 'react'
-import { makeStyles } from '@material-ui/styles'
+import { makeStyles } from '@mui/styles'
 
 const useStyles = makeStyles(
   {
