@@ -1,7 +1,7 @@
 import React, { memo } from 'react'
 import { useSelector } from 'react-redux'
-import { AppBar, Container, Toolbar } from '@material-ui/core'
-import { makeStyles } from '@material-ui/styles'
+import { AppBar, Container, Toolbar } from '@mui/material'
+import { makeStyles } from '@mui/styles'
 import Menu from './components/Menu'
 import Side from './components/Side'
 import Fab from './components/Fab'
