@@ -1,7 +1,8 @@
 import React from 'react'
 import { useSelector } from 'react-redux'
-import { StylesProvider, ThemeProvider } from '@material-ui/styles'
-import { CssBaseline } from '@material-ui/core'
+import { StyledEngineProvider, ThemeProvider } from '@mui/material/styles'
+import { StylesProvider } from '@mui/styles'
+import { CssBaseline } from '@mui/material'
 import './global.css'
 import * as themes from './themes'
 import { generateClassName } from '../libraries/styles'
@@ -11,11 +12,13 @@ export default ({ children }) => {
   const theme = light ? themes.light : themes.dark
 
   return (
-    <StylesProvider generateClassName={generateClassName}>
-      <ThemeProvider theme={theme}>
-        <CssBaseline />
-        {children}
-      </ThemeProvider>
-    </StylesProvider>
+    <StyledEngineProvider injectFirst>
+      <StylesProvider generateClassName={generateClassName}>
+        <ThemeProvider theme={theme}>
+          <CssBaseline />
+          {children}
+        </ThemeProvider>
+      </StylesProvider>
+    </StyledEngineProvider>
   )
 }
